Remove 'Sec*' prefix from all type names
Security Framework adds a leading "Sec" prefix to all type and function
names to namespace them.

This makes sense given it's a C library, but is a bit redundant in a
Rust binding.

The underlying types and functions wrapped are meticulously documented,
so we don't really gain much by mimicing this naming scheme in a
binding.